requirements.txt
git+https://github.com/dolfies/discord.py-self.git
aiohttp[speedups]

# Raspberry Pi optimizations
psutil
gevent

# DNS fallback support
dnspython